Refactor rule and rule set construction
Move `build_ruleset` to `rules.py` to keep all related logic grouped together.
Also refactor _make_tiff_metadata_rule into a class function, making the class
responsible for generating the objects. _rule_function_mapping and make_rule
have been rolled together into _build_rule.
